infra_auth_stack/services/requirements.txt
# Lambda Layer Dependencies
# These packages will be installed in the Lambda execution environment

# JWT token handling
PyJWT==2.8.0

# HTTP requests to existing APIs
requests==2.31.0

# AWS SDK (already available in Lambda, but specified for completeness)
boto3==1.34.0
